Update test file for advancing runners
For one player on base for each hit (single, double, triple, homerun) test
result to see if base runner advances appropriatley.
