--- test/test01.py
@@ -1,2 +1,0 @@
-def testA():
-    assert (1, 2, 3) == (1, 2, 3)
